fix history layout
