+++ app/components/Faucet.tsx
@@ -1,6 +1,21 @@
 "use client";
 
-import { SimpleGrid, Center, Box, Input, Stack, Text } from "@chakra-ui/react";
+import {
+  SimpleGrid,
+  Center,
+  Box,
+  Input,
+  Stack,
+  Text,
+  Heading,
+  Spinner,
+  Tbody,
+  Td,
+  Th,
+  Thead,
+  Tr,
+  Table,
+} from "@chakra-ui/react";
 import Image from "next/image";
 import { ChangeEvent, useState, useEffect } from "react";
 import { GitHub } from "./Github";
@@ -8,6 +23,16 @@ import { SendButton } from "./SendButton";
 import { Balance } from "./Balance";
 import { Telegram } from "./Telegram";
 import { Twitter } from "./Twitter";
+import useSWR from "swr";
+import dayjs from "dayjs";
+import { sanitizeAddress } from "../api/utils";
+import Link from "next/link";
+import { HistoryRate } from "./HistoryRate";
+import { get_history_blockscout } from "../api/tables";
+import relativeTime from "dayjs/plugin/relativeTime";
+
+
+dayjs.extend(relativeTime);
 
 export const Faucet = () => {
   const [walletAddress, setWalletAddress] = useState("");
@@ -23,7 +48,15 @@ export const Faucet = () => {
     localStorage.setItem("walletAddress", walletAddress);
   }
 
+
+const { data, error, isLoading } = useSWR("/api/history", () =>
+  get_history_blockscout(8, walletAddress)
+);
+
+console.log("API Response:", data);
+
   return (
+    <>
     <Box
       display="flex"
       flexDirection="column"
@@ -75,5 +108,86 @@ export const Faucet = () => {
         </SimpleGrid>
       </Stack>
     </Box>
+    <Box padding="20px" maxHeight="100vh" overflow="auto">
+      <Heading size="lg" marginBottom="20px">
+        Transfer History
+      </Heading>
+      {isLoading && <Spinner />}
+      {error && (
+        <Text color="red.500">Error loading data: {error.message}</Text>
+      )}
+      {!isLoading && !error && data && (
+        <>
+          {Array.isArray(data) ? (
+            <Table variant="simple">
+              <Thead>
+                <Tr>
+                  <Th>Txid</Th>
+                  <Th>Value</Th>
+                  <Th>Timestamp</Th>
+                  <Th>Status</Th>
+                </Tr>
+              </Thead>
+              <Tbody>
+                {data.map((item, index) => (
+                  <TransferRow key={index} {...item} />
+                ))}
+              </Tbody>
+            </Table>
+          ) : data.items && Array.isArray(data.items) ? (
+            <Table variant="simple">
+              <Thead>
+                <Tr>
+                  <Th>Txid</Th>
+                  <Th>Value</Th>
+                  <Th>Timestamp</Th>
+                  <Th>Status</Th>
+                </Tr>
+              </Thead>
+              <Tbody>
+                {data.items.map((item, index) => (
+                  <TransferRow key={index} {...item} />
+                ))}
+              </Tbody>
+            </Table>
+          ) : (
+            <Text>
+              No transfer history available or unexpected data format.
+            </Text>
+          )}
+        </>
+      )}
+      <HistoryRate />
+    </Box>
+    </>
+  );
+};
+
+
+const TransferRow = (props: {
+  hash: string;
+  value: string;
+  timestamp: string;
+  status: string;
+}) => {
+  const address = props?.hash || "Unknown";
+  const url = `https://scan.exsat.network/address/${address}`;
+  const short = sanitizeAddress(address);
+  const time = props.timestamp ? dayjs(props.timestamp).fromNow() : "Unknown";
+  const value = props.value
+    ? (BigInt(props.value) / BigInt(10 ** 18)).toString()
+    : "Unknown";
+
+  return (
+    <Tr>
+      <Td>
+        <Link href={url} target="_blank" rel="noreferrer">
+          {short}
+        </Link>
+      </Td>
+      <Td>{value} BTC</Td>
+      <Td>{time}</Td>
+      <Td>{props.status || "Unknown"}</Td>
+    </Tr>
   );
 };
