+++ app/page.tsx
@@ -5,7 +5,7 @@ import Navbar from "./components/Navbar/Navbar";
 import { GoogleAnalytics } from "nextjs-google-analytics";
 import { ChakraProvider } from "@chakra-ui/react";
 import { Container, Flex } from "@chakra-ui/react";
-import { TransferHistory } from "./components/History";
+
 import { Faucet } from "./components/Faucet";
 import Footer from "./components/Footer/Footer";
 import { RecoilRoot } from "recoil";
@@ -25,7 +25,7 @@ export default function Home() {
             alignItems="center"
           >
             <Faucet />
-            <TransferHistory />
+           
           </Flex>
         </Container>
         <Footer />
